Removed daniels changes. It broke my backend code.
Added CORS to the hosted pages.

Edited `dockerRun.sh` to allow for it to run with password and only on my computer.

Dependency changes in pom.xml to allow for hot reload as gradle, as it annoyed my that maven did not allow hot reloads.

Lastly, added put request method, so now we have implemented CRUD in full